Task3/requirements.txt
streamlit
google-generativeai
langdetect
deep-translator